permisos y final del proyecto 1.0

+++ VeterinarioSW/ControlDeVeterinaria/views.py
@@ -1,16 +1,50 @@
 from django.shortcuts import render, redirect
-from .models import Propietario, Mascota, PagoServicio, Cita, Producto, PagoProducto, Proveedor, Comprador, Factura, Cirugias, Enfermedades, Otros, Vacunas
-from .forms import PropietarioForm, MascotaForm, PagoServicioForm, CitaForm, ProductoForm, PagoProductoForm, ProveedorForm, CompradorForm, FacturaForm, CirugiasForm, EnfermedadesForm, OtrosForm, VacunasForm
+from .models import Propietario, Mascota, PagoServicio, Cita, Producto, PagoProducto, Proveedor, Comprador, Factura, Cirugias, Enfermedades, Otros, Vacunas, Account
+from .forms import PropietarioForm, MascotaForm, PagoServicioForm, CitaForm, ProductoForm, PagoProductoForm, ProveedorForm, CompradorForm, FacturaForm, CirugiasForm, EnfermedadesForm, OtrosForm, VacunasForm,AccountForm
 from django.views.generic import (UpdateView)
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
+from django.contrib.auth import get_user_model
+from datetime import date
+
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
-def login(request):
+def loginUser(request):
+
+    if request.method == "POST":
+        usernameCompare = request.POST.get('username')
+        passwordCompare = request.POST.get('password')
+        usuarioEncontrado= False
+
+        User = get_user_model()
+        users = User.objects.all()
+        for user in users:
+            if user.username == usernameCompare and user.password == passwordCompare:
+                usuarioEncontrado = True
+        
+        if usuarioEncontrado:
+            login(request, user)
+            return render(request, 'ControlDeVeterinaria/index.html')     
+        else:
+            messages.info(request, 'Nombre de usuario o contraseña erroneo')
     return render(request, 'ControlDeVeterinaria/log.html', {})
 
+def logoutUser(request):
+    logout(request)
+    return redirect('/')
+
+@login_required(login_url='login')
 def index(request):
     return render(request, 'ControlDeVeterinaria/index.html', {})
 
+def calendario(request, mes=1):
+    dateHoy=date.today()
+    print(mes)
+    return render(request, 'ControlDeVeterinaria/calendario.html', {})
+
+@login_required(login_url='login')
 def verMascota(request):
     listaMascotas = Mascota.objects.all()
 
@@ -19,6 +53,7 @@ def verMascota(request):
     }
     return render(request, 'ControlDeVeterinaria/mascota.html', context)
 
+@login_required(login_url='login')
 def crearMascota(request):
     form = MascotaForm(request.POST)
     if form.is_valid():
@@ -29,11 +64,13 @@ def crearMascota(request):
     }
     return render(request, 'ControlDeVeterinaria/crearMascota.html', context)
 
+@login_required(login_url='login')
 def borrarMascota(request,id):
     Mascota.objects.get(id=id).delete()
 
     return redirect('/mascota/')
 
+@login_required(login_url='login')
 def crearPropietario(request):
     listaPropietario = Propietario.objects.all()
     form = PropietarioForm(request.POST)
@@ -45,11 +82,13 @@ def crearPropietario(request):
     }
     return render(request, 'ControlDeVeterinaria/crearPropietario.html', context)
 
+@login_required(login_url='login')
 def borrarPropietario(request,id):
     Propietario.objects.get(id=id).delete()
 
     return redirect('/propietario/')
 
+@login_required(login_url='login')
 def expediente(request):
     formCirugias = CirugiasForm(request.POST)
     formVacunas = VacunasForm(request.POST)
@@ -71,6 +110,7 @@ def expediente(request):
     }
     return render(request, 'ControlDeVeterinaria/expediente.html', context)
 
+@login_required(login_url='login')
 def expediente_mascota(request, id):
     cirugias = Cirugias.objects.filter(mascota__id=id)
     vacunas = Vacunas.objects.filter(mascota__id=id)
@@ -88,6 +128,7 @@ def expediente_mascota(request, id):
     }
     return render(request, 'ControlDeVeterinaria/expediente-mascota.html', context)
 
+@login_required(login_url='login')
 def crearCita(request):
     form = CitaForm(request.POST)
     listaCitas = Cita.objects.all()
@@ -99,10 +140,12 @@ def crearCita(request):
     }
     return render(request, 'ControlDeVeterinaria/citas.html', context)
 
+@login_required(login_url='login')
 def borrarCita(request,id):
     Cita.objects.get(id=id).delete()
     return redirect('/citas/')
 
+@login_required(login_url='login')
 def pagoServicio(request):
     form = PagoServicioForm(request.POST)
     listaPagos = PagoServicio.objects.all()
@@ -114,10 +157,12 @@ def pagoServicio(request):
     }
     return render(request, 'ControlDeVeterinaria/pagos.html', context)
 
+@login_required(login_url='login')
 def borrarPago(request,id):
     PagoServicio.objects.get(id=id).delete()
     return redirect('/pagos/')
 
+@login_required(login_url='login')
 def producto(request):
     form = ProductoForm(request.POST)
     listaProductos = Producto.objects.all()
@@ -129,10 +174,12 @@ def producto(request):
     }
     return render(request, 'ControlDeVeterinaria/productos.html', context)
 
+@login_required(login_url='login')
 def borrarProducto(request,id):
     Producto.objects.get(id=id).delete()
     return redirect('/productos/')
 
+@login_required(login_url='login')
 def pagoProducto(request):
     form = PagoProductoForm(request.POST)
     listaPagoProductos = PagoProducto.objects.all()
@@ -144,11 +191,16 @@ def pagoProducto(request):
     }
     return render(request, 'ControlDeVeterinaria/pago_productos.html', context)
 
+@login_required(login_url='login')
 def borrarPagoProducto(request,id):
     PagoProducto.objects.get(id=id).delete()
     return redirect('/pago-productos/')
 
+@login_required(login_url='login')
 def factura(request):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     form = FacturaForm(request.POST)
     listaFactura = Factura.objects.all()
     if form.is_valid():
@@ -159,18 +211,30 @@ def factura(request):
     }
     return render(request, 'ControlDeVeterinaria/factura.html', context)
 
+@login_required(login_url='login')
 def borrarFactura(request,id):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     Factura.objects.get(id=id).delete()
     return redirect('/factura/')
 
+@login_required(login_url='login')
 def verFactura(request,id):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     datosFactura = Factura.objects.get(id=id)
     context = {
         'datosFactura':datosFactura,
     }
     return render(request, 'ControlDeVeterinaria/ver_factura.html', context)
 
+@login_required(login_url='login')
 def proveedor(request):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     form = ProveedorForm(request.POST)
     listaProveedor = Proveedor.objects.all()
     if form.is_valid():
@@ -181,18 +245,30 @@ def proveedor(request):
     }
     return render(request, 'ControlDeVeterinaria/proveedor.html', context)
 
+@login_required(login_url='login')
 def borrarProveedor(request,id):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     Proveedor.objects.get(id=id).delete()
     return redirect('/proveedor/')
 
+@login_required(login_url='login')
 def verProveedor(request,id):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     datosProveedor = Proveedor.objects.get(id=id)
     context = {
         'datosProveedor':datosProveedor,
     }
     return render(request, 'ControlDeVeterinaria/ver_proveedor.html', context)
 
+@login_required(login_url='login')
 def comprador(request):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     form = CompradorForm(request.POST)
     listaComprador = Comprador.objects.all()
     if form.is_valid():
@@ -203,13 +279,44 @@ def comprador(request):
     }
     return render(request, 'ControlDeVeterinaria/comprador.html', context)
 
+@login_required(login_url='login')
 def borrarComprador(request,id):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     Comprador.objects.get(id=id).delete()
     return redirect('/comprador/')
 
+@login_required(login_url='login')
 def verComprador(request,id):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
     datosComprador = Comprador.objects.get(id=id)
     context = {
         'datosComprador':datosComprador,
     }
     return render(request, 'ControlDeVeterinaria/ver_comprador.html', context)
+
+@login_required(login_url='login')
+def crearAccount(request):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
+    form = AccountForm(request.POST)
+    listaAccounts = Account.objects.all()
+    if form.is_valid():
+        form.save()    
+    context ={
+        'listaAccounts':listaAccounts,
+        'form':form,
+    }
+    return render(request, 'ControlDeVeterinaria/accounts.html', context)
+
+@login_required(login_url='login')
+def borrarAccount(request, id):
+    rol = request.user.rol
+    if rol == 'EMPLEADO':
+        return redirect('/index/')
+    Account.objects.get(id=id).delete()
+    return redirect('/accounts/')
